+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Text optimization in the import log
+- Optimization of the last PHPStan warnings
+
 ## [5.0.1] - 27.10.2025
 
 ### Changed